Question4/shahbaz_project/shahbaz_project/task_project/requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bootstrap4==0.1.0
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.4
cryptography==2.9.2
distlib==0.3.1
Django==3.1.7
django-filter==2.4.0
django-multifilefield==1.0
django-multiselectfield==0.1.12
djangorestframework==3.12.2
dnspython==1.16.0
Faker==5.1.0
filelock==3.0.12
httpie==2.4.0
idna==2.10
importlib-metadata==3.7.2
isort==5.7.0
lazy-object-proxy==1.4.3
lxml==4.6.2
Markdown==3.3.4
mccabe==0.6.1
multidict==4.7.6
mysql-client==0.0.1
mysql-connector-python==8.0.19
mysqlclient==2.0.3
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.2.0
Pillow==7.1.2
pipenv==2020.11.15
protobuf==3.6.1
psycopg2==2.8.6
pycparser==2.20
Pygments==2.8.1
pylint==2.6.0
PyMySQL==0.9.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.24.0
requests-toolbelt==0.9.1
scapy==2.4.4
selenium==3.141.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.4.2
virtualenv-clone==0.5.4
webdriver-assistant==0.0.7
webdriver-controller==0.0.4
websockets==8.1
wrapt==1.12.1
xmltodict==0.12.0
xvfbwrapper==0.2.9
yarl==1.5.1
zipp==3.4.1
